+++ tox.ini
@@ -4,7 +4,7 @@ envlist = py3{10,11,12}
 min_version = 4.0
 
 [testenv]
-whitelist_externals = poetry
+allowlist_externals = poetry
 commands =
     poetry install -v
     poetry run pytest tests/
